added Depth First Search
